+++ resources/views/livewire/base-crud/base-crud.blade.php
@@ -325,6 +325,8 @@
 
     @include('ptah::livewire.base-crud.partials._modal-delete')
 
+    @include('ptah::livewire.base-crud.partials._modal-import')
+
     {{-- Loading overlay apenas para ações pesadas (salvar, deletar, exportar) --}}
     <div wire:loading.delay.long wire:target="save,deleteRecord,export"
         class="fixed inset-0 z-40 flex items-center justify-center bg-black/20">