+++ src/Livewire/BaseCrud/Concerns/HasCrudForm.php
@@ -668,6 +668,7 @@ trait HasCrudForm
             'delete' => ['showDeleteButton', 'delete'],
             'restore' => ['showTrashButton', 'restore'],
             'export' => [null, 'export'],
+            'import' => [null, 'import'],
             default => [null, null],
         };